fix(schema): version attribute of Function::S3Location in SAM is optional
SAM specification marks this field as optional but is marked as required
in this repo.
SAM spec is here - https://github.com/awslabs/serverless-application-model/blob/master/versions/2016-10-31.md#s3-location-object

See generate/sam-2016-10-31.json for the change. The rest of the changes
come from running `go generate`.

fixes #87

+++ cloudformation/resources/aws-cognito-userpool_admincreateuserconfig.go
@@ -19,7 +19,7 @@ type AWSCognitoUserPool_AdminCreateUserConfig struct {
 	// UnusedAccountValidityDays AWS CloudFormation Property
 	// Required: false
 	// See: http://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-cognito-userpool-admincreateuserconfig.html#cfn-cognito-userpool-admincreateuserconfig-unusedaccountvaliditydays
-	UnusedAccountValidityDays float64 `json:"UnusedAccountValidityDays,omitempty"`
+	UnusedAccountValidityDays int `json:"UnusedAccountValidityDays,omitempty"`
 
 	// _deletionPolicy represents a CloudFormation DeletionPolicy
 	_deletionPolicy policies.DeletionPolicy
